move class to the only place it is used
