+++ eval/CMakeLists.txt
@@ -80,5 +80,4 @@ vespa_define_module(
     src/vespa/eval/instruction
     src/vespa/eval/onnx
     src/vespa/eval/streamed
-    src/vespa/eval/tensor/dense
 )
