+++ searchlib/src/tests/tensor/dense_tensor_store/dense_tensor_store_test.cpp
@@ -8,7 +8,6 @@ LOG_SETUP("dense_tensor_store_test");
 #include <vespa/eval/eval/value.h>
 #include <vespa/eval/eval/value_type.h>
 #include <vespa/eval/eval/test/value_compare.h>
-#include <vespa/eval/tensor/dense/mutable_dense_tensor_view.h>
 
 using search::tensor::DenseTensorStore;
 using vespalib::eval::SimpleValue;
