+++ searchlib/src/vespa/searchlib/features/dense_tensor_attribute_executor.h
@@ -4,7 +4,7 @@
 
 #include <vespa/searchlib/fef/featureexecutor.h>
 #include <vespa/eval/eval/value.h>
-#include <vespa/eval/tensor/dense/mutable_dense_tensor_view.h>
+#include "mutable_dense_tensor_view.h"
 
 namespace search::tensor { class ITensorAttribute; }
 namespace search::features {
